Handle multi-line strings in HTML properly
Change handling of strings from lang/en/tool_crawler.php which seem to
be on several lines on purpose and which do not contain HTML line
breaks. Make these strings HTML-ready for output, that is, replace the
line breaks with '<br>' each, and HTML-escape the strings.

This is done under the assumption that the two affected strings
(‘recentactivitydesc’ and ‘urldetails_help’) should not contain HTML
snippets.