(1/2) Switch test string to translatable string
Do not use a hardcoded string when printing out the
“Hello robot: '<username>'” message on the test page. Instead, use the
appropriate language string there in order to match the code in
crawler::is_bot_valid which does the string comparison.

The comparison code was written the wrong way around: the string printed
on the test page had not been internationalised, while the expected output
string had been. This commit results in both of them internationalised.

NB: as a side effect, this requires a future code change to also ensure
that the comparison is done _with the bot locale_ and not with the user
locale!

The nice effect for users is that the test page can be translated and
users see the translation (even though a test page is of limited value
to users).