(2/2) Correctly compare hello string in bot test
The crawler self-test queries a test page, which prints strings
according to the current locale. When the crawler code runs a self-test,
it uses its scraping function on the test page and compares the result
to an expected string.

Ensure that the expected string is looked up _with the bot locale_ and
not with the locale that the current admin user has set. This makes the
locale settings independent: the bot can use any locale (or the site
default locale), and admin users can, too.

The test page can now be translated in any language and users will see
the translation (even though a test page is of limited value to them),
and the bot self-test will still work.