Do not store fixed-language string in database
When libcurl returned an error at scrape time, a string with an
(English) error message was composed and stored in the database. It was
retrieved from there to be shown when a user looked at a report.

This behaviour has to be changed in order to internationalise the code.
The database should never store translated (or English) strings. Storing
such strings is not acceptable because
* a string lookup at scrape time would necessarily use the bot locale
  (or a fixed locale/fixed strings), but
* the bot locale may (and, in some installations, most likely will)
  differ from the user’s locale for the user viewing a report, and
* a translation (at report time) of a string composed at scrape time is
  unreliable as long as the string is not prepared for unambiguous
  parsing _and_ as messages from different sources and for different
  cases could possibly clash (see below).

A better way to handle the error case is to have the database store that
part of the error message which cannot (or should not) be translated.
The complete string to be shown to the user should be composed when it
is needed (and it should be in the user’s locale, of course).

So when a report is generated for a user, it must be possible to
distinguish what happened at scrape time:
* either a failure was reported by libcurl
* or libcurl returned successful completion of the request, but a remote
  web server signalled an error.

It was not possible to reliably detect (from looking at rows in the URI
table in the database) whether libcurl at crawl time had been successful
or had failed. Using the value from the "httpcode" column or parsing the
value from the "httpmsg" column is not reliable, as both may overlap
with legitimate answers from remote web servers (even though this is
highly unlikely and only probable in cases where remote server operators
intend to disturb our crawler).

The problems are solved by introducing a new column ("errormsg") which
stores error messages from the HTTP user agent (in this case, libcurl).

The unreliable process above could still be seen as an acceptable way
during database upgrade (if one really liked to have the content
upgraded): "httpcode" must equal '500' in this case, *and* "httpmsg"
must start with 'Curl Error: ' and be followed by an integer. This is
not taken care of during the database upgrade in this commit because it
is assumed that users rather have the bot re-crawl their sites anyway in
order to get this issue solved.

"errormsg" purposefully is a column with type _text_: it is intended to
store information returned about an error from the HTTP user agent
library. In the existing implementation, this is always libcurl, but
could in theory be any HTTP user agent, which might return error strings
instead of integer values. This may be an oversubtle differentiation,
and could easily be worked around by a future database upgrade, but then
again, it works quite fine.

curl_errno returns an integer value, and with libcurl as HTTP user
agent, it is this value which is stored as a string in column
"errormsg". A transition to, for example, the return value of curl_errno
(which is a string) is possible. That latter value is currently put into
the column for the page title, and this commit avoids stepping out of
line with regards to this part of data storage of the existing
implementation.

Future extensions could use introduce a new "errorcode" column for
integer return values (like from curl_errno), in addition to "errormsg"
for text about the error (error message, like from curl_error). In case
of an error from the HTTP user agent, "title" would be left as NULL. In
case of HTTP user agent success, "errorcode" and "errormsg" would both
be NULL. It remains to be seen, though, why this would be *far* better
than the solution introduced in this commit.

The test cases are adjusted just for consistency: "errormsg" is added to
new objects/database entries, but the existing test cases require the
property to be set to NULL anyway.

To sum up, this commit does two things:
1. improves crawler internationalisation by making a constant string
   localisable,
2. increases the quality of stored scrape results by allowing to find
   out _at report-generation time_ whether an error _at scrape time_ had
   been reported by the HTTP user agent (libcurl) or by a remote web
   server.