Remove more dead code
The code which is being removed is (or rather, was) responsible for
excluding links based on the exclude configuration settings. It is
already included in function mark_for_crawl, where it is better suited.
Function mark_for_crawl is called by function link_from_node_to_url,
which again is called by function parse_html – the function from which
the code is dropped.

Removal does not make any difference (except saving some CPU cycles),
because the code computes a value for variable $bad which is never used
afterwards.