Fix order of newline replacing and HTML escaping
Commit 3c728d07bcc18ff24da5e3248debd68599b6a408 in pull request #57
erroneously had two lines in the wrong order, twice. And it
unnecessarily used htmlentities instead of htmlspecialchars.

The correct way is to call htmlspecialchars, which is sufficient, and to
first replace HTML special characters, and afterwards insert HTML markup
into the string (by replacing newline characters with '<br>').

htmlspecialchars is invoked in HTML 4.01 mode so that newline characters
are not replaced by the '&NewLine;' character reference (which is
specific to HTML 5).