(1/3) Unit testing: generate URL page in string
Use a string variable for storing the generated url detail pages, and
output all content at the end of the user-facing script.

This is done in preparation for further changes which should make it
easier to write unit tests for the pages.

This may have performance implications, as the string may be copied over
and over again (I have not checked the PHP/Zend engine implementation of
the .= operator). Using an array with page chunks could help here. But
as the expected frequency of accesses to the URL detail page is very
low, this is fine.

The changes are only related to unit testing (like with PHPUnit), and
not to behavioural testing (like with Behat).