Fix wrong link text/page title usage after trim
When a link text (content of the a element) or a page title (content of
the title element) in a handled HTML document is '0' after trimming, it
was treated as empty. This is wrong.

trim($string) always returns a string. The expression (!'0') evaluates
to TRUE. So the result of trim can and should be handled differently.

As a result, link texts or page titles like '0', '  0   ' and so on
work.