Properly unescape/escape document titles
Fix issues with handling of linked document titles:

1) the titles extracted from the DOM of the link targets (if they are
   HTML documents) may still contain HTML entity references. We need to
   remove them before storing the document titles in the database.

   In case a link title contains a broken entity reference, simply
   ignore it/handle it like PHP does (which may result in ugly output,
   but that is none of our concern, as the linked document already is in
   error).

2) the titles need HTML-escaping before they are written to the output
   which is sent to our user. This is not strictly dependent on 1), and
   is needed in any case, because we want to produce clean HTML output.

This commit changes function tool_crawler_link, and adds a new parameter
with a default value. It does not modify the way things work for callers
of this function which want to link to things other than document
titles.

TODO: database update function which does 1) on all document titles
which are already stored in the database.