Do not query DOM for document title twice
We should not call potentially expensive operations on the DOM if we can
avoid it. It also makes the code a bit easier to follow.

This results in a minor performance improvement (and a minor increase in
memory consumption) during the runtime of function parse_html.

The isset call has been changed from the (auto-generated) property to
the node object on purpose, after verifying that this works properly.