Security fix: use purified HTML a element content
Do not copy the link content to the database, but only text nodes, and
even further clean the content up by dropping some potential child
nodes. This is necessary to prevent scripting or other attacks by
prepared links inside or outside of Moodle installations.

Without this commit, HTML elements like script, iframe, img, object, and
so on can be included as link text, and the crawler would store them in
the database and later display them on report pages. Reports can be
shown to course admins, but also to Moodle admins.

Even keeping images as link content is not recommended, as keeping them
allows for the following attack vectors:
1) accesses to remote servers when someone watches a crawler report,
2) phishing-like attacks where an image looks like a piece of the Moodle
   web user interface,
3) attacks that put high pressure on a graphical web client by including
   images with immensely big resolutions.

Fixes the most security-related part of #65.