Apply HTML escaping on link text before output
Ensure that link text from the database (where the plain text is stored,
without HTML escapes) is properly HTML-escaped when it is used in HTML
output.

In report.php, function tool_crawler_link is used with the link text.
Even though it now has an argument that allows to have the function
HTML-escape its argument, do not make use of that. This provides the
foundation for better processing in future for the cases when there is
no link text: look up the localized string with the message, HTML-escape
it, and afterwards add a bit of HTML markup to mark it accordingly. The
corresponding places are marked with a comment.