Harden reading and escape on writing of HTML IDs
The HTML standard explains how to handle ID attributes: almost
everything is allowed but spaces.

We need to protect from the case that an HTML document has something
like 'id="foo #bar"' in its markup, as this would cause clashes on our
side. In principle, handling such things is a task for the HTML library,
but it returns the plain ID value to us.

The case where id="0" must be allowed, so change the code accordingly.

We must decode HTML entities before checking the IDs, because this is
also not done by the HTML library. This piece of code is therefore
in line with previous commits handling attribute values from the DOM.

Before the values from the database are sent to the web client, HTML
escaping must be applied again. So the unescaping/escaping part of this
commit is in line with previous commits handling values from HTML
documents/for HTML output.