HTML-escape headings
The headings are constant localised strings. It could be useful to allow
HTML markup in them. Nevertheless, this commit ensures HTML-escaping is
performed on the strings so that language string maintainers do not have
to think about the strings (you cannot see from the strings whether they
are meant to be HTML snippets or not).

This commit should be reverted (and language strings switched) as soon
as there is a real necessity to use HTML markup in headings (e.g. for
emphasizing a word).