Escaping in error message on test page redirection
If the Hello Robot test page issues a redirection, the final URI is
included in the error message. The corresponding function returns an
HTML snippet if it returns an error. In the case of a strange server
misconfiguration on our side, and after multiple redirections, the final
URI might be out of our control. Anyway, it is a non-HTML string which
is output to HTML, so we should HTML-escape it before output.