Delete cookie jar on plugin uninstallation
Make use of an uninstall script to delete the cookie jar when the plugin
is uninstalled. This ensures that we do not leave the file lying around.
A failure to delete the cookie jar is silently ignored.

It is unnecessary to keep the cookie jar for the future in case the
plugin is installed again, as new cookies will be created and written to
the jar.