Use the reason-phrase from the final HTTP response
When bot queries are redirected (server replies with 3xx code), the full
reply obtained from curl contains several response headers and
status-lines. Without this commit, the _final_ HTTP status-code is
obtained from curl, but the _first_ reason-phrase is parsed from the
reply.

This commit changes the response headers parsing code so that the
_final_ HTTP reason-phrase is used instead.

Fixes #36.
Fixes #37.

As a side effect, the information about the type of redirects is not
visible anymore to the user. This problem could be tackled in future by
a modification in the crawler reports: they could show which kind of
redirection has happened. In addition to that, all hops in a redirection
chain could be stored when crawling the links. This would allow to also
show the hops in the reports.

For the time being, it is more difficult for users to employ measures
in case of permanent redirects (that is, update links that cause
permanent redirections), simply because the users can not see from a
report whether a redirect is permanent or not.