Use a keep-alive connection on scraping
I can not see why a fresh connection should be forced. Disabling this,
as it will also aid when we first send a HEAD and then (sometimes) a GET
request in future.

According to the PHP documentation, ‘Connection: keep-alive’ is the
default; not verified. Setting the option explicitly does not harm, and
we are safe.