Use HEAD request first when scraping
When a link is followed, do not use HTTP GET directly, but first try
with HTTP HEAD. Resources do not have to be downloaded with GET at all
when the following conditions are met:
* the remote server understands and handles HEAD,
* the size of the document is known, and
* the resource is not an HTML document.

HTML documents have to be downloaded so that their title can be
extracted.

While this increases the number of HTTP requests, it will usually save
resources on server side and on the network: a HEAD request and its
response only need a few bytes, but asking a server with GET may consume
some more server resources. If the response to a GET request is fully
sent over the network, this can unnecessarily transfer a big amount of
data.