Enforce a download size limit
The transfer callback function aborts the download when the newly
introduced size limit is reached. This limit is a hard-coded reasonable
value and does not have to be configured by the user. It is chosen in a
way that it will normally contain the document title in the case of HTML
documents.