Use a higher size limit on redirections
This is necessary so that the curl redirection logic can be used, but
transfers to non-big resources are not aborted prematurely if the
targets are behind a redirection (with a big HTTP body).

Note: even though curl does not return intermediate documents on
redirections (it only returns all headers and the final document body),
the second download limit is required. This is because curl may still
call the progress function while downloading (without returning) bodies
of redirecting HTTP responses.