Extract check for external URI into function
The code checks several times whether an URI is external. Extract all
the checks into a common function, unifying and even improving them a
bit.

The new function returns boolean, and its return value is directly
stored in the node object for the link. That object used to store an
integer, 0 or 1, for the external flag before. The new behaviour is not
a problem because the database functions convert a boolean to an integer
before writing the value to the database.

This commit changes the logic in one very special case: if a link to an
_external_ document is crawled, but the external URI redirects to a page
_inside_ the Moodle installation, the link is now treated as internal.
This means that the HTML page will be fully parsed: the course
identifier and the like will be extracted, and links will be followed.