Use write callback function to retrieve data
We plan to introduce code changes that will abort the download if the
target resource is too large. Even if the target is an overlarge
external HTML document, we can still make use of its title. But
curl_exec does not return the partial content if the download has been
aborted. So store the downloaded data in a way that always makes it
available, no matter whether curl has signalled an error or not.

We temporarily need lots of memory when the downloaded chunks are
concatenated. However, this way of storing and converting the data in
one go is much more efficient than appending to a string each time
(because this produces a new string object in PHP).