Remove unused fields from URL page database query
This saves resources, especially because we do not have to JOIN with the
course table.

The code is only executed when the URL information page, url.php, is
requested. It is not a bad idea to include course short names with the
links presented there, but the existing code was dead. It can be brought
back if and when the URL page is extended in future.