Improve code readability around a variable
After employing the write callback function (see earlier commit), the
call to libcurl no longer returns the response data, but a boolean value
instead. It signals whether the operation was successful (true) or not
(false).

Use an appropriately named variable and test its value in a more
commonly used way. This makes the code easier to understand.