Move header handling to a header callback function
We do not want to implement complicated HTTP response parsing if we can
avoid it. So we use curl to do the hard work.

Ensure that the write callback function is only called for response
_body_ content. As curl handles redirections and does not pass bodies of
redirecting resources to that function, the write callback function will
only be called with (zero to many) octets that belong to the target
response body. We are going to make use of that later in order to detect
whether our web client has reached the body of the HTTP response for the
target resource. For now, nothing changes.

This is implemented by two changes: first, have curl no longer return
header content in the write callback function. Second, introduce a new
header callback function, and shift all header processing to it.