Start with original URIs after redirections
When a GET request is issued after a HEAD request that has been
redirected at least once, the request URI had been internally set to the
final URI by libcurl code or Curl functions in PHP. It is correct to use
this value in the case of (only) permanent redirects. But if at least
one temporary redirect is reached, this becomes wrong.

In order to work around this, we would have to track the last URI. Then
we can determine the last URI that was a target of the last permanent
redirect before the first temporary redirect (if any). Do not implement
this for now, as the web may be too broken anyway for it to work
reliably.

While a waste of network resources and time, restarting at the original
URI should always work, even in case of chains of permanent
redirections. This case will probably not be found very often.