Store resource size also in case of an error
Always store the resource size. If it is unknown, store that we do not
know it (value -1). In cases where the result of a HEAD request is being
handled, the value may be overwritten later after a subsequent GET
request.

Ensures that the resource size, which is useful information, is always
available if possible, even if there has been an error during
downloading.