Allow storing details about the resource size
When a link is crawled, it depends on the remote system and on local
configuration (which is going to be introduced by changes in near
future) whether the crawler fully downloads a resource or aborts the
download in between. As a result, the resource size can be known, be
unknown, or we can know that it is at least the number of bytes
downloaded so far, but we cannot tell more about it. If the server has
not sent a header field with the resource size, the latter two cases are
common.

Add a column to the URI table in the database. This column indicates
details about the stored size: it can give the relation to the actual
resource size, or can tell that the actual resource size is unknown.

Add constants for the possible values in this database column. Having
these constants in a separate file ensures that we do not keep copied
definitions synchronized between the upgrade script (db/upgrade.php) and
other parts of the code. The constants are require()d from lib.php in
order to make them publicly available; for now, the code in lib.php does
not need them. Other constants are not moved because they are just a
means of keeping the code clean from magic numbers, but do not need to
be publicly available.

Provide a clean upgrade path: up to now, the crawler code has always
determined exact resource sizes (or NULL, on errors). So set this as the
default value during database upgrade.