Use size details in oversize decision
If we do not know a resource size at all, consider it “maybe big” and
include it in the oversize statistics (report, link to it, history): we
cannot tell how big the resource is, so it may be big.

Do not include queued links in the big files report if they have not yet
been processed because only links which have already been scraped can be
big.

Make resources which are _at least_ a certain size appear before
resources which are _exactly_ this size in the oversize report because
this resembles a kind of natural ordering that users will probably
expect.