Allow localizations of symbol for “unknown”
A question mark could be used in two places (for HTTP code and resource
size) in order to indicate that the respective value is not known.
Change that to ensure that 100 % of the plugin strings can be
translated. This now allows to use a different character than a question
mark, depending on the locale.