Allow localisation of symbol for “unknown”
A question mark has been used in place of the HTTP status-line if it
shall be indicated that the value is not known. Change that to a
localisable string to ensure that 100 % of the plugin strings can be
translated. This now allows to use a different character than a question
mark, depending on the locale.