Dedicated function to determine the resource size
Rewrite file size decisions in order to utilize as much useful
information as possible. Take into account the Content-Length header if
it is available, and the number of bytes already downloaded if
applicable.

If servers send wrong Content-Lengths, the stored resource sizes will be
wrong, too, and reports may incorrectly list the “big” status of
documents.

The biggest benefit is that the new code can set
TOOL_CRAWLER_FILESIZE_ATLEAST. This allows for more exact reports.

Replace the magic value -1.0 for “size unknown/unusable” with a superior
implementation.