Implement first two network strain levels
“reasonable” and “resolute” (or higher) settings are differentiated now.
This means that, if not operating in “reasonable” mode, the crawler will
abort GET requests for non-HTML documents, no matter whether their
length is known from the HTTP header or not. In everyday use, this
should not be an issue because most non-HTML resources are static
anyway, and the servers know their size in advance and put it in
Content-Length.

If not working in “reasonable” mode, the download size limit is the
configured big file size. If servers send a Content-Length, this will
not cause big downloads. If they do not, this is a good thing, as it is
the proper way to determine whether the document is actually larger than
this size.