Implement “excessive” and “wasteful” modes
If the network strain setting is set to “excessive”, try to determine
the exact length of HTML documents at all cost, and GET the full
document so that it is guaranteed that a title can be parsed. This is
far too much in all reasonable cases.

In “wasteful” mode, even do this if the length of the document is known
from the headers. It should never be necessary to enable this at all,
except perhaps when testing the server delivering the document.

Non-HTML resources are currently not affected: their download is aborted
if the size is known, as the crawler could not use their content anyway.