Add configuration option to disable the bot
When the option is enabled, the scheduled tasks will do nothing.

This option can be useful, for example, if you like to disable resource
consumption by the crawler on your development systems, and like to do
this by means of a configuration setting instead of by disabling the
scheduled tasks.