Reset external status on re-scrapes if necessary
When a link is scraped for the second time or later, make sure that its
external flag is correctly updated in all cases. If the linked
(internal) resource originally redirected to another (external)
resource, but the redirect has gone in the meantime, this resulted in a
problem: the status had been set to ‚external‘ and it had kept this
value.

Now the status is always updated, no matter whether we are processing a
redirection or not. This always leads to correct results.