Ensure correct document titles after recrawls
The database is always updated when we crawl a link. Scraping only
changes the title in case of an error. HTML parsing only happens if
there is HTML to parse. And then, the HTML document might even have no
title. The situation may have changed when a link is crawled repeatedly.
So it is necessary to start with no title each time a link is crawled.
The title will of course still be properly set if necessary, but the
code change ensures that it is also properly unset if it should _not_ be
set.