Ensure correct HTTP status-line after re-scraping
If scraping a link for the repeated time, it may be necessary to store a
new HTTP status-line. As the link error status prevents the stored HTTP
status-line from being used if it would not have been updated, there has
not been a problem in displayed reports. But the database entries could
have become partially (and unneccessarily) confusing, which is fixed by
this change.