Uglify code to mute code checker
The moodle_local_codechecker code has some questionable quirks. For
example, it seems not to get along with some ways of line-breaking.

So, in order to get the codechecker tests in Travis runs to behave,
sacrifice readability of the code for green instead of red lights. ☹

It would be a good idea to revert this after moodle_local_codechecker
has been improved.